// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number;
  title: string;
  slug: string;
  layout: (
    | {
        heading: string;
        subheading: string;
        image: number | Media;
        cta_button: {
          label: string;
          url: string;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        heading: string;
        content: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'content';
      }
    | {
        heading: string;
        form: number | Form;
        id?: string | null;
        blockName?: string | null;
        blockType: 'newsletter-form';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}